chore: digits params layout + refactoring
