+++ packages/trader/src/AppV2/Components/BottomNav/bottom-nav.tsx
@@ -17,9 +17,10 @@ import { useHistory, useLocation } from 'react-router';
 type BottomNavProps = {
     children: React.ReactNode;
     className?: string;
+    onScroll?: (e: React.UIEvent<HTMLDivElement>) => void;
 };
 
-const BottomNav = observer(({ children, className }: BottomNavProps) => {
+const BottomNav = observer(({ children, className, onScroll }: BottomNavProps) => {
     const history = useHistory();
     const location = useLocation();
     const { active_positions_count, onMount, onUnmount } = useStore().portfolio;
@@ -92,7 +93,9 @@ const BottomNav = observer(({ children, className }: BottomNavProps) => {
     }, []);
     return (
         <div className={classNames('bottom-nav', className)}>
-            <div className='bottom-nav-selection'>{children}</div>
+            <div className='bottom-nav-selection' onScroll={onScroll}>
+                {children}
+            </div>
             <div className='bottom-nav-container'>
                 {bottomNavItems.map((item, index) => (
                     <BottomNavItem
