+++ packages/trader/src/AppV2/Components/TradeParameters/LastDigitPrediction/last-digit-prediction.scss
@@ -5,6 +5,7 @@
     padding: var(--core-spacing-400) var(--core-spacing-800);
     border-radius: var(--core-spacing-400);
     background: var(--core-color-opacity-black-75);
+    width: 100%;
 
     &__title {
         color: var(--core-color-opacity-black-600);
